requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
Werkzeug==2.2.3
email-validator==2.0.0
Pillow==9.5.0
python-dotenv==1.0.0
WTForms==3.0.1
SQLAlchemy==2.0.15
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
markupsafe==2.1.2
# AI libraries
textblob==0.17.1
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.1
joblib==1.2.0
